backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.31
pydantic==2.7.3
anthropic==0.29.0
PyMuPDF==1.24.7
python-docx==1.1.2
python-dotenv==1.0.1
aiofiles==23.2.1
